requirements.txt
# Gmail Watcher Dependencies
# Install with: pip install -r requirements.txt

google-api-python-client>=2.80.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-auth>=2.25.0

# LinkedIn Posting Dependencies
python-dotenv>=1.0.0
playwright>=1.40.0
